+++ R/aaa.R
@@ -1,0 +1,5 @@
+# aaa.R - Constants, etc.
+# These are defined first and generally not exported
+
+  pkg = 'cache.sodium'
+  extension = 'sodium.rds'