+++ R/utils_rds.R
@@ -1,0 +1,6 @@
+write_rds <- function( object, path, ... ) {
+  saveRDS( object, file=path, ... )
+  invisible(object)
+}
+
+read_rds <- readRDS