+++ R/zzz.R
@@ -1,7 +1,7 @@
 .onAttach <- function( libname, pkgname ) {
 
-  # suppressWarnings( try( v <- utils::packageVersion(pkgname, libname), silent = TRUE ))
-  # version <- if( exists('v') ) paste0("-", v ) else ""
+  suppressWarnings( try( v <- utils::packageVersion(pkgname, libname), silent = TRUE ))
+  version <- if( exists('v') ) paste0("-", v ) else ""
 
   dt <- read.dcf( system.file("DESCRIPTION", package = pkgname ), "Date" )
   yr <- substr(dt,1,4)
@@ -15,9 +15,11 @@
       , " - Copyright \u00a9 ", ifelse(! is.na(yr), yr, '')
       , " Decision Patterns"
       , domain = NA
-    )
+  )
 
-  if( require(future) )future::plan(multiprocess)
+  # Register this package.
+  if( require(cache) ) cache_register_sodium()
+  if( is.null(cache_backend() ) ) cache_use_sodium()
 
 }