+++ NEWS.md
@@ -1,7 +1,8 @@
 # cache.sodium 0.1.0
 
-* Refactor cache.sodium
-* Added a `NEWS.md` file to track changes to the package.
+ * Use `cache::cache_register_backend()`
+ * Refactor cache.sodium
+ * Added a `NEWS.md` file to track changes to the package.